+++ artifacts/kuhni-na-zakaz/.next/static/chunks/webpack.js
@@ -190,7 +190,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("c3cf4cd0e7170e61")
+/******/ 		__webpack_require__.h = () => ("0dd03f74ecaf1a8a")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */